rearranging files and allowing for diverse entanglement tests
